Moved the update logic thread to separate class to avoid repeated code
