+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2020
+  ~ Paketin Seuranta
+  ~
+  ~ @author developerfromjokela
+  ~ @author norkator
+  -->
+
 <resources>
 
     <!-- Not translatable -->
@@ -262,6 +270,7 @@
     <string name="settings_tweaks">Tweaks</string>
     <string name="parcel_phase">Phase</string>
     <string name="location_not_available">Location not available</string>
+    <string name="notification_message">Parcel\' status changed: %1$s - %2$s</string>
 
 
     <!-- Strings related to Settings -->
